pyproject.toml
[tool.poetry]
name = "etl-video-encoding"
version = "0.1.0"
description = ""
authors = ["pert <petr777@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.10"
prefect = "^2.6.1"
ffmpeg-python = "^0.2.0"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
aiobotocore = "^2.4.0"
alembic = "^1.8.1"


[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
